Fetch version numbers and copyright year from Git
Instead of committing a change for each release to update the release
number for the application and provider, use the output from git describe
(minus the commit hash for the actual version, but full output stored in
informational version).  Also use the date of the most recent commit to
set the 'to' copyright year in the assemblies.  This is a tidied up and
expanded version of https://gist.github.com/thoemmi/3483648.