Provider handling class refactoring and cleanup
Bring together provider related code from the Data class and the static
plugin class into a new Provider class.  This contains both the static
initialization code to find the providers and the properties and methods
needed to store and expose the provider data and instance creation.  This
also allows access to the provider class, name, description, icon and
options handler without having to create an instance of the provider each
time one of these is required.  Also, store the loaded assemblies rather
than the path to them as this is cleaner and should improve performance.