Get SDK path from the registry for build scripts
Get the path to the Windows SDK from the
HKLM\SOFTWARE\Microsoft\Microsoft SDKs\Windows\v7.1 key in the registry
instead of having it hard-coded in the build and publish batch files.

This fixes the scripts looking in Program Files (x86) instead of Program
Files when being run from a 32-bit command prompt.