Require provider user-agent for CachedWebClient
Update the signature of DownloadData and DownloadString in CachedWebClient
to force non-internal callers (e.g. providers) to pass a user-agent string
to be sent along with the request.  This allows requests from different
providers to be separately identified on servers and should hopefully
prevent Radio Downloader from being tarred with the same brush as a rogue
provider.

Also updated the Podcast provider to pass a sensible user agent for requests
that it makes.