Check databases are not corrupt at startup
After opening connections to the template database, the user database
and the search index database (all at startup), execute a pragma
integrity_check command to test if the sqlite database is valid or has
become corrupted.

If the template or user databases are corrupt then refuse to start and
show a message to the user (in the case of the user db with an option to
open the help page about this).  With the search index, just delete and
re-create it automatically.

Resolves #191.