Add NOT NULL constraints to columns in episodeext
If a piece of episode ExtInfo data was returned from the provider with
a value of null, this was stored in the database, causing a crash when
it was loaded back again.  To prevent this from occurring, add a NOT
NULL constraint to the value column in the episodeext table.

Also add NOT NULL to the epid and name columns for completeness as
this is not inferred by SQLite for primary key columns (although these
are unlikely to be set to null in practice).