Handle lines in podcast summary same as html desc
Move the multiple blank line collapsing code out of the HtmlToText
class into a function in the main PodcastProvider and also apply it to
text from the <itunes:summary> tag.  This ensures that all podcast and
episode descriptions are processed in the same way.

Also convert CRLF to LF to ensure that the collapsing code works
consistently.