Only invalidate underneath tabs when on Windows
On Windows it is necessary to invalidate the section of the tab bar
underneath the buttons to ensure that it gets re-painted correctly and
gives the illusion that the active tab continues to the edge of the
control.  However, under linux this is unnecessary and causes the rest
of the form to never be painted, so only call Invalidate on Windows.