Add the option to view the feed URL for podcasts
It is useful to be able to view the URL for subscribed or favourited
feeds to check if they are up to date.  Implemented this by adding a
method to return a delegate for showing extra programme information and
a 'More Info' button on the toolbar which is shown if the provider for
that programme supports it.

Fixes #201.