Enable additional built in code analysis warnings
Adjust the built in code analysis warnings to flag names or member
access that could be simplified and to warn about unnecessary casts.
Fix the violations found by these, but suppress them for form Designer
files and generated resource designer files.

Also completely disable the 'Remove this Qualification' warning as this
conflicts with StyleCop.