Add a specific hintpath for 64-bit System.Web
For some reason the build process picks up the x86 version of
System.Web.dll when building for x64 which causes a warning and the
build to fail (as all warnings are converted to errors).  To work around
this, set a specific HintPath pointing to the correct location for x64
builds.

Also re-enable win64 builds in the AppVeyor matrix as this should now
actually build successfully.