Fix providers not showing under Mono on Linux
The Type.GetInterface() call under Mono appears to be buggy if the
optional second parameter (ignoreCase) is set to true as it erroneously
returns false for matching interface names.  This causes no provider
classes to be found in the provider assemblies.

As the interface name that we are passing uses the correct case anyway,
just use the case-sensitive path which works correctly.