Correctly handle titles with slashes in filenames
Backslashes (under Windows) and forward-slashes (non-Windows) are
allowed in the filename template so that a directory structure can be
created.  However, they are not handled in programme or episode titles
which causes unexpected directories to be created.

To fix this, replace instances of Path.DirectorySeparatorChar in
programme and episode names with spaces before assembling the full path.