Move signtool actions from batch files to MSBuild
Add new 'Signed' configuration (with the same settings as Release), and
an MSBuild target which is trigged by the Signed configuration to run
signtool against the generated binaries and installer for the Windows
platforms.

Remove the build batch files as these are no-longer useful and simplify
the publish batch file.  Also update the publish batch file to work with
Visual Studio 2015.