mono: Don't abort before sending the error report
Due to slight behaviour differences, calling Environment.Exit from the
FormClosing event of the error reporting form causes the application to
terminate before sending the report under Mono.

Instead, move the Environment.Exit to the unhandled exception handling
method as this works with both .NET and Mono (and is also a bit
cleaner).