mono: Move useDefaultCredentials fix to instance
790b7bbbc0 resolved exceptions which occurred on any web request (due to
Mono not understanding useDefaultCredentials for a proxy), by clearing
the global default proxy in the CachedWebClient constructor.  However,
this doesn't resolve non-cached requests if they are made before the
first cached request (e.g. an unhandled exception on startup).  To
resolve this, move the workaround to the AppInstance.OnCreateMainForm
method.