Overhaul date stripping from episode titles
Add a great deal more tests for stripping dates from episode titles with
a much greater range of formats and styles.  Overhaul the
StripDateFromName() function in the TextUtils class to pass all these 
tests.  As part of rewriting the regex, make it much more maintainable 
by splitting it up into components which are re-assembled at
compile-time into the final expression.

Also enhance the functionality which only removes the date if it is 
similar to the date of the episode: instead of just matching on the 
month and year, parse the matched date and see if it is within 6 days of
the episode date before removing it.