Refactor the provider API to use an abstract class
Instead of exposing the provider API via an interface, switch to
providing an abstract class which can be inherited from to implement a
provider.  This allows having optional properties / methods / events
which can have a stub implementation in the abstract class and only need
to be replaced if needed, saving boilerplate code in the providers.

As part of the refactoring, also move the provider API related code into
a separate RadioDld.Provider namespace to tidy up the clutter of classes
in the base RadioDld namespace and rename the Providers class to Handler
as this is a better description of what it does.