Refactor CachedWebClient slightly for unit testing
Instead of providers fetching the CachedWebClient instance via a public
static method, make an instance available as a RadioProvider property.
Also split CachedWebClient into an abstract base class and an
implementation so that a test class can easily be created for unit
testing and refactor out the static GetInstance() function.

Add a new project for testing the podcast provider and add a basic test
with basic podcast details.