Consume and produce Simple Chapters data
Initial implementation of consuming Podlove Simple Chapters data in the
Podcast Provider and producing it (if available) in the generated RSS
feed.

This is a fairly wide-ranging enhancement as it required support for
chapters to be added to the provider interface, database and model.

As chapter data may only be available after a download is complete, add
optional chapter information to the data returned after a successful
download.

Also add a test for returning chapter information from the podcast
provider, using the Podlove example feed data.