Don't treat a cancelled download as completed
Add a return statement after handling the cancellation of a download so
that it isn't treated as completed before deletion.  This fixes a crash
trying to build a final filename.

The logic error was introduced in 17dc36bf (clean download cancellation)
and was promoted into a crash by d841a4dc (simple chapters handling) as
it attempts to dereference the null DownloadInfo object from the Podcast
Provider.

Resolves #225.