Switch toolbar layout table to autosize
As part of the improvements to the Windows Forms support in Mono 5.14
and above, a bug seems to have been introduced which causes a hang if we
set the height of TableToolbars at Form.Load for the main form.

Instead, set TableToolbars to autosize as this achieves the same
end-result but without the hang at startup under Mono.