Start making use of Null-conditional operator ?.
The Null-conditional operator ?. allows us to clean up quite a lot of
the code which is wrapped in checks for null.  This also improves the
thread-safety of things like event handlers as ?. internally uses a
local variable.

Also re-enable showing IDE1005 as a warning now that the code triggering
it has been replaced.