Package for Arch Linux as well as Windows
To support packaging for both Windows and Linux, rename the Installer
project to 'Package Windows' and add a 'Package Linux' project which
executes a new package-linux bash script to do the packaging.

To start with, just package with fpm for Arch Linux.  Add a locally
built binary of Microsoft.VisualBasic.dll as this is a dependency and
isn't available in a standard Arch package.

Also adjust appveyor.yml to install the packaging dependencies and
retain the built package as a build artifact.