Gracefully handle vacuum failing with space error
Vacuuming the database will fail if there is less disk space than the
size of the database in either the SQLite temporary directory or the
same directory as the database.

As the database can get quite large, vacuum isn't that critical to
performance and this prevents starting the application, catch an
SQLiteException with an error code of 'Full' and postpone the next
vacuum attempt by 7 days.

Fixes #232.