Fix main window being disabled after Status shown
Due to a bug in Mono, e.CloseReason is set to CloseReason.UserClosing in
Status_FormClosing() when Hide() is called.  This causes e.Cancel to be
set to true which prevents the Mono code from re-enabling the parent
forms correctly.

To work around this issue, only set e.Cancel to true in
Status_FormClosing() when running under Windows.

Resolves #234.