Fix an sqlite3.dll build warning shown in VS
Since 547b44000d, an warning is shown after each build in Visual Studio
saying that it cannot add a link to sqlite3.dll as it is within the
project directory tree.  This doesn't break the build but is unsightly.

Switch from using <Content> and <Link> for the sqlite3.dll copy to
<ContentWithTargetPath> to fix this warning.