Report all podcast download error details to user
Currently, only DNS resolution, HTTP 403 and HTTP 404 errors are caught
by the Podcast Provider during the download process and reported to the
user, any other download issue is reported as an 'Unknown Error'.

To simplify the provider and greatly improve the information provided
back to the user, set the DownloadException type based on the
WebException type and set the error details to the message of the
innermost exception.

Resolves #235.