Store images in db without recompressing as PNG
Currently images are returned from the provider as System.Drawing.Bitmap
objects.  This makes them simple to display but inefficient to store in
the database.  Images are converted to PNG before storage (as of
97ddda0565), but as the majority of podcast images are JPEG this causes
a considerable size increase for no benefit.

Address this issue by adding a new 'CompressedImage' class and using
this to pass the original compressed image data back from the provider
for direct storage in the database.

To try and ensure that invalid image files don't get stored, check in
the CompressedImage ctor that the data can be loaded into Bitmap
(although bypass this check for performance when fetching images back
from the database).

CompressedImage has properties 'Width', 'Height' and 'Image' to make
using the images nearly as simple as before.

Resolves #228.