Change Episode & Programme GetImage() to property
Having to pass the episode id or programme id to the Programe and
Episode static GetImage() methods is a bit clumsy.

Replace them with properties as this is both cleaner and saves a
database query for each call as the image id can be fetched as part of
the initial class details query.