Overhaul podcast image retrieval and add tests
The RSSNodeImage function which fetches podcast programme and episode
images is functionally unchanged since 096dde5bfc (in 2008!) and has
some suboptimal behaviour including silently ignoring any exceptions
which occur.  It also uses the CreateNamespaceMgr function to handle
namespaces which means it requires them to be defined on the root <rss>
node and use the prefixes of 'itunes' and 'media' instead of checking
the namespace name itself.

Replace the RSSNodeImage function with a new FetchImage function which
takes a parent XML node, xpath query & namespace manager and returns
either the retrieved image or null.  This allows the Programme and
Episode code to look in different places for images: <image>/<url> is
not valid in an <item> node and <media:thumbnail> is hardly used for
<channel> images.

Replace the odd 'dynamic' namespace handling code in CreateNamespaceMgr
with correct definitions of the iTunes and Media RSS namespaces.

The existing Chapter Image tests have been extended to include Programme
and Episode image tests which check namespace handling, image precedence
and a number of failure scenarios.