Perform releases directly from AppVeyor
Now that releases are no-longer signed and that installers are stored as
artifacts from the build jobs on AppVeyor, update appveyor.yml to
automatically create a draft GitHub release with the built artifacts
attached instead of building them locally.

Also update the release process to reflect this change and remove the
old batch files used for the Windows builds.