Fix StripDateFromName missing dates after a number
StripDateFromName doesn't remove a date if it is proceeded by a number
in certain cases.  This is due to the number, day and month being
matched as day, month, year and then either being an invalid date or not
being close enough to the episode date.

Add and adjust tests to cater for this issue and resolve it by checking
the rest of the string for matches if a pattern matched but wasn't a
valid date or close enough to remove.