Use InternalsVisibleTo instead of public for tests
Use the InternalsVisibleTo attribute in the main application and podcast
provider for their respective test assemblies.  This allows classes
previously changed to public just to allow testing to be changed back to
internal and should require less access modifier changes for additional
tests in future.

This does unfortunately break the CA1812 and CA1823 checks so disable
them for the time being.