Target the .NET Framework 4.5.2 instead of 2.0
Windows 7 will not be supported by Microsoft beyond the 14 January so it
is a good time to move from targeting .NET 2.0 to .NET 4.x. Supported
versions of Windows will no-longer include .NET 2.0 - 3.5 by default but
will include .NET 4.5.1 or above.

As version 4.5.1 of .NET exited support in January 2016, target 4.5.2 as
this is defined as an OS component by Microsoft and will be supported
for as long as the underlying OS.

To support this change, the following adjustments needed to be made:
 - Stop marking assemblies as security transparent.  This was added in
   1be4a048c0 to satisfy the now deprecated CA2122 FxCop rule but causes
   issues under .NET 4.7.2.
 - Update from xUnit 1.9.2 to 2.4.1 to work with .NET 4.7.2.
 - Adjust the .NET framework Condition in the Windows installers to
   check for 4.7.2 or above using the WiX NetFxExtension.