Remove the OsUtils.TrayAnimate function
This was added a long time ago to give an animation to show that the
application had minimized to the system tray.  The Windows API call that
it uses for the animation (DrawAnimatedRects) does nothing under Windows
Vista and above when composition (DWM) is enabled.  As Windows 8 and
above don't allow composition to be disabled the animation is never
displayed in currently supported versions of Windows.

There have been error reports recently from a number of users where
TrayAnimate has failed to find the system tray location, so as this
function is no-longer doing anything useful just remove it.