Build filename at end of download instead of start
Currently the downloaded file name generation takes place at the start
of each download.  This works mostly okay with the default naming
template but if the pattern is changed to something less unique and
multiple downloads happen at once the download can fail right at the end
due to an already existing file.

Adjust the provider interface so that the final filename is no-longer
passed to the DownloadProgramme function and instead add a 'Downloaded'
property to DownloadInfo in which the downloaded file can be returned as
a TempFileBase.  This allows the final filename to only be generated at
the point of moving the file.