Add basic .editorconfig and use just lf for *.cs
Add an .editorconfig file to set the indent size & style and final
newline handling to match the settings in stylecop.json and set charset
to utf-8.

Also take the opportunity now that end_of_line is supported by Visual
Studio to set this to lf (as well as adjusting .gitattributes to match)
for consistency across all platforms.