Fix exception infinite loop in MoveToSaveFolder
Some of the other exceptions that File.Move raises are subclasses of
IOException (which we catch as plain IOException is raised for the
destination file existing) but as adjusting the filename doesn't resolve
these other exceptions the current handling can get stuck in an infinite
loop.  Test that what we have caught is IOException rather than a
subclass before continuing.