Windows: Work around MAX_PATH for downloaded files
The .NET framework doesn't cope well with paths longer than MAX_PATH - 1
(259 characters) due to the limitations of the underlying Win32 API
calls it uses.

To work around this limitation when saving downloaded files to a path
longer than MAX_PATH under Windows use P/Invoke to call MoveFileW with
the special 'long path' prefix (which is not subject to the MAX_PATH
limit).

It looks like the P/Invoke won't be necessary once we are targeting .NET
4.6.2 or above and the 'long path' prefix won't be necessary once we are
targeting .NET Core.

Resolves #239.