Filter podcast episodes when finding them by ID
If a podcast feed contains an <item> node without a valid title or
enclosure URL but that has an ID which is the same as a later <item>
which is valid, the Podcast Provider will match the episode ID to the
first invalid node instead of the later valid one.

Add a test for this situation and then resolve the problem by applying
the same filtering when finding an episode by ID as when building the
list of available episodes.

Fixes #241.