Throw better exceptions in OsUtils.MoveFile
The current implementation of OsUtils.MoveFile throws a Win32Exception
for any errors that occur but this behaviour differs from .NET
File.Move which it is replacing which causes the code that calls it to
handle the exceptions incorrectly.

Add a test and fix the behaviour when the destination file exists, the
destination folder doesn't exist or the destination filename is invalid.