Move HttpListener.Start to RssServer.Start
RssServer.Start is called by the main form even if starting the
HttpListener instance failed in the RssServer ctor, leading to an
unhandled ObjectDisposedException.

To resolve this, move the HttpListener.Start call into RssServer.Start,
so that the HttpListener.BeginGetContext call can be skipped if the
listener cannot be started.

Fixes #246.