Fix crash if RSS server hits a protocol conflict
If the port that the RSS server is configured to listen on has a URL
reservation with a different protocol than http (https for example), the
application currently crashes, meaning that the user cannot change the
port to one without a conflicting protocol.

Resolve this by showing a messagebox with details of the failure and
continue running without the RSS server.

Resolves #245.