Move author copyright to individual files
Instead of trying to handle the list of authors separately to the file
copyright, remove AUTHORS.md and change the copyright in all the files
to the contributor names (this effectively reverts e222497235).  This
helps to keep the author copyright with the file content if any of the
files are used elsewhere.

While updating the copyright headers, also correct the 'from' copyright
date in a number of files which was incorrectly set to 2007 in all files
in 7e0a09f069.

Also take the opportunity to reduce the size of the file headers by
removing the GPL boilerplate from each file and replacing it with a
SPDX-License-Identifier header line.