Show details when request for a new podcast fails
Currently, if a WebException is thrown when trying to request the feed
for a new podcast, a fixed error message is displayed ("There was a
problem requesting the feed from the specified URL") which is clear but
a little vague.  To make this more useful, append the exception message
(which with WebException is generally quite informative).  Hopefully
this will shed a little more light on issue #250.

Also do the same for XmlException in this situation although this is
probably less useful.