Prevent download hang if filename exceeds fs limit
If the name template ends up generating a path where the filename itself
exceeds the limits of the file system (e.g. 255 characters for NTFS on
Windows, 255 bytes for most Linux filesystems), the try..catch block to
handle race-condition file moves in MoveToSaveFolder unhelpfully ends up
causing an infinite loop.

Add a test for the correct behaviour of an IOException being thrown and
adjust MoveToSaveFolder to check for the existence of a file in the
destination path if an IOException has been caught, re-throwing if it
has been thrown for a different reason.

Also adjust the error message displayed for a download which has failed
at this point with an IOException to include the possibility of the
file name template needing to be adjusted.

Resolves issue #259.